refactor(AuthenticationFormContainer): Simplify component by moving form and Google login button to separate module
